==> code/6-13-bbs_did_graph.Rmd <==
---
title: "6-13-bbs_did_graph"
author: "Zeyu Lou"
date: "2024-06-13"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r packages}
rm(list=ls())

library('foreign')
library('haven')
library(readxl)
library(dplyr)
library(ggplot2)
library(plm)
```

## 1. bird biodiversity did

#### 1.0 Import data
```{r import}
mdp_data <- read_xlsx("../output_data/MDP_1980_2008_updated.xlsx")
county_x_year_panel_data <- read.csv("../output_data/county_x_year_panel_data.csv")
```

#### 1.1 merge the data
```{r merge}
# format fips as integer in both dataset
mdp_data <- mdp_data %>%
  mutate(fips = as.integer(fips),
         plant_id = paste0(Number, Winner, fips, `City (If available)`)) # create a unique id for different plants
n_distinct(mdp_data$plant_id) # 310
n_distinct(mdp_data$fips) # 224


# Loop through each fips in mdp_data to check if it exists in county_x_year_panel_data
unique_fips_mdp <- unique(mdp_data$fips)
unique_fips_county <- unique(county_x_year_panel_data$fips)
not_found_count <- 0
not_found_ls <- c()
for (fips in unique_fips_mdp) {
  if (!(fips %in% unique_fips_county)) {
    not_found_count <- not_found_count + 1
    not_found_ls <- c(not_found_ls, fips)
  }
}
not_found_count # 81. 
not_found_ls
filtered_mdp_data <- mdp_data %>% filter(!(fips %in% not_found_ls)) 
nrow(filtered_mdp_data) # 197

# merge the dataset if they have the same fips
merged_mdp_data <- mdp_data %>%
  left_join(county_x_year_panel_data, by = "fips")

# Since the year has no missing values, we want to delete the data that does not have a corresponding FIPS code.
merged_mdp_data <- merged_mdp_data %>%
  filter(!is.na(`year`))

n_distinct(merged_mdp_data$plant_id) # 197. verified

```

#### 1.2 filter the years

```{r}
mdp_data_10yrs <- merged_mdp_data %>% # mdp_data_year
  mutate(`Publication Year` = as.integer(`Publication Year`),
         year = as.integer(year),
         year_diff = year - `Publication Year`,
         post_treatment = ifelse(year_diff > 0, 1, 0),
         winner_indicator = ifelse(Winner == 1, 1, 0),
         runner_up_indicator = ifelse(Winner == 0, 1, 0)) %>%
  filter(year_diff >= -4 & year_diff <= 5)

n_distinct(mdp_data_10yrs$plant_id) # 170

# mdp_data_10yrs <- mdp_data_10yrs %>%
#   group_by(plant_id) %>%
#   filter(n() == 10) %>%
#   ungroup()

n_distinct(mdp_data_10yrs$plant_id) # 85
table(mdp_data_10yrs$Winner) # 0:490; 1:360
```


#### 1.3 did plots

```{r}
mdp_data_10yrs <- mdp_data_10yrs %>%
  mutate(log_abundance = log(abundance),
         log_richness = log(richness))

mdp_data_10yrs.1 <- pdata.frame(mdp_data_10yrs, index = c("County", "year"))

formula_abundance <- log_abundance ~ winner_indicator * post_treatment + runner_up_indicator * post_treatment + factor(Number) + factor(County) + factor(year)
formula_richness <- log_richness ~ winner_indicator * post_treatment + runner_up_indicator * post_treatment + factor(Number) + factor(County) + factor(year)

model_abundance <- plm(formula_abundance, data = mdp_data_10yrs, model = "within")
model_richness <- plm(formula_richness, data = mdp_data_10yrs, model = "within")

summary(model_abundance)
summary(model_richness)


# 提取预测值
mdp_data_10yrs$predicted_abundance <- predict(model_abundance)
mdp_data_10yrs$predicted_richness <- predict(model_richness)

# 提取残差
mdp_data_10yrs$residuals_abundance <- residuals(model_abundance)
mdp_data_10yrs$residuals_richness <- residuals(model_richness)

# 计算每个年份和组别的平均值
avg_data <- mdp_data_10yrs %>%
  group_by(year_diff, winner_indicator) %>%
  summarize(mean_abundance = mean(predicted_abundance, na.rm = TRUE),
            mean_richness = mean(predicted_richness, na.rm = TRUE))

avg_data$winner_indicator <- factor(avg_data$winner_indicator, levels = c(0, 1), labels = c("Runner-up", "Winner"))


bird_abundance_did <- ggplot(avg_data, aes(x = year_diff, y = mean_abundance, color = winner_indicator, linetype = winner_indicator)) +
  geom_line(size = 1) +
  labs(title = "Bird Abundance (log)", x = "Year relative to plant opening year", y = "Log Abundance") +
  scale_color_manual(values = c("Winner" = "blue", "Runner-up" = "orange")) +
  scale_linetype_manual(values = c("Winner" = "solid", "Runner-up" = "dashed")) +
  theme_minimal()

bird_richness_did <- ggplot(avg_data, aes(x = year_diff, y = mean_richness, color = winner_indicator, linetype = winner_indicator)) +
  geom_line(size = 1) +
  labs(title = "Bird Richness (log)", x = "Year relative to plant opening year", y = "Log Richness") +
  scale_color_manual(values = c("Winner" = "blue", "Runner-up" = "orange")) +
  scale_linetype_manual(values = c("Winner" = "solid", "Runner-up" = "dashed")) +
  theme_minimal()

bird_abundance_did
bird_richness_did

# plots data visualization
bird_abundance_visual <- ggplot(mdp_data_10yrs, aes(x = year_diff, y = log_abundance, color = as.factor(Winner))) +
  geom_line(stat = "summary", fun = mean) +
  labs(title = "Log Abundance over Time",
       x = "Year Difference",
       y = "Log Abundance",
       color = "Winner") +
  scale_x_continuous(breaks = seq(-4, 5, by = 1))+
  theme_minimal()

bird_richness_visual <- ggplot(mdp_data_10yrs, aes(x = year_diff, y = log_richness, color = as.factor(Winner))) +
  geom_line(stat = "summary", fun = mean) +
  labs(title = "Log Richness over Time",
       x = "Year Difference",
       y = "Log Richness",
       color = "Winner") +
  scale_x_continuous(breaks = seq(-4, 5, by = 1))+
  theme_minimal()

bird_abundance_visual
bird_richness_visual




```

## 2. tree biodiversity did

#### 2.1 import data
```{r import tree}
# tree_data <- readRDS("../data/species count data/result.rds")
# write.csv(tree_data, "../data/result.csv")
```

#### 2.2 merge data
```{r merge tree}
# tree_data <- tree_data %>%
#   mutate(
#     STATECD = sprintf("%02d", STATECD),
#     COUNTYCD = sprintf("%03d", COUNTYCD),
#     fips = paste0(STATECD, COUNTYCD)
#   )
# 
# tree_data_fips <- tree_data %>%
#   group_by(fips, MEASYEAR) %>%
#   summarize(
#     abundance = sum(county_count, na.rm = TRUE),
#     richness = n_distinct(SPCD)
#   ) %>%
#   ungroup()

# write.csv(tree_data_fips, "../output_data/tree_data.csv")

tree_data_fips <- read.csv("../output_data/tree_data.csv")

n_distinct(tree_data_fips$fips) # 3114

tree_data_fips <- tree_data_fips %>%
  mutate(fips = as.integer(fips))

tree_mdp_data <- mdp_data %>%
  left_join(tree_data_fips, by = "fips")

# # verifying steps.
# unique_fips_mdp <- unique(mdp_data$fips)
# unique_fips_county <- unique(tree_data_fips$fips)
# not_found_count <- 0
# not_found_ls <- c()
# for (fips in unique_fips_mdp) {
#   if (!(fips %in% unique_fips_county)) {
#     not_found_count <- not_found_count + 1
#     not_found_ls <- c(not_found_ls, fips)
#   }
# }
# not_found_count # 10
# not_found_ls

# tree_mdp_data.test <- tree_mdp_data %>% filter(!is.na(MEASYEAR))
# n_distinct(tree_mdp_data.test$fips) # 214: 224-10=214. verified!
# n_distinct(tree_mdp_data.test$plant_id) # 292

# create year_diff as x_axis

tree_mdp_data_10yrs <- tree_mdp_data %>%
  mutate(`Publication Year` = as.integer(`Publication Year`),
         MEASYEAR = as.integer(MEASYEAR),
         year_diff = MEASYEAR - `Publication Year`) %>%
  filter(year_diff >= -4 & year_diff <= 5)

```

#### 2.3 did plots

```{r}

tree_mdp_data_10yrs <- tree_mdp_data_10yrs %>%
  group_by(plant_id) %>%
  filter(n() == 10) %>%
  ungroup() %>%
  mutate(log_abundance = log(abundance),
         log_richness = log(richness))

n_distinct(tree_mdp_data_10yrs$plant_id) # 15

tree_abundance_did <- ggplot(tree_mdp_data_10yrs, aes(x = year_diff, y = log_abundance, color = as.factor(Winner))) +
  geom_line(stat = "summary", fun = mean) +
  labs(title = "Log Abundance over Time",
       x = "Year Difference",
       y = "Log Abundance",
       color = "Winner") +
  scale_x_continuous(breaks = seq(-4, 5, by = 1))+
  theme_minimal()

tree_richness_did <- ggplot(tree_mdp_data_10yrs, aes(x = year_diff, y = log_richness, color = as.factor(Winner))) +
  geom_line(stat = "summary", fun = mean) +
  labs(title = "Log Richness over Time",
       x = "Year Difference",
       y = "Log Richness",
       color = "Winner") +
  scale_x_continuous(breaks = seq(-4, 5, by = 1))+
  theme_minimal()

tree_abundance_did
tree_richness_did

```
